Fix pet petted and bowl filled checks
Seems like 'wasPetToday' was removed as a reflection property of the Pet
in SDV 1.4. Dug around and found that .grantedFriendshipForPet does
effectively the same thing.

Also, not sure if this was a bug or not, but I did find that you can
access _farm.petBowlWattered instead of manually probing the water bowl
building. Guessing that this will help custom maps work!

This commit was tested on PC for both single- and multi-player and
appears to function correctly.